# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API за интерактивна карта на социалните предприятия в Североизточен район
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Enterprise {
  id: number;
  naimenovanie: string;
  /** @nullable */
  eik?: string | null;
  oblast: string;
  grad: string;
  /** @nullable */
  adres?: string | null;
  lat: number;
  lng: number;
  /** Клас А или Клас А+ */
  klas: string;
  pravnaForma: string;
  ikonomicheskaDeynost: string;
  celevnaGrupa: string;
  socialnaKauza: string;
  /** @nullable */
  socialnaInovaciya?: string | null;
  kratkoOpisanie: string;
  /** @nullable */
  misiya?: string | null;
  /** @nullable */
  telefon?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  producti?: string | null;
  /** @nullable */
  uslugi?: string | null;
  /** @nullable */
  broyZaeti?: number | null;
  /** @nullable */
  broyUyazvimiLica?: number | null;
  /** @nullable */
  godinaZastoyvane?: number | null;
  /** @nullable */
  istoriyaNaUspeha?: string | null;
}

export interface EnterpriseInput {
  naimenovanie: string;
  eik?: string;
  oblast: string;
  grad: string;
  adres?: string;
  lat: number;
  lng: number;
  klas: string;
  pravnaForma: string;
  ikonomicheskaDeynost: string;
  celevnaGrupa: string;
  socialnaKauza: string;
  socialnaInovaciya?: string;
  kratkoOpisanie: string;
  misiya?: string;
  telefon?: string;
  email?: string;
  website?: string;
  facebook?: string;
  logoUrl?: string;
  producti?: string;
  uslugi?: string;
  broyZaeti?: number;
  broyUyazvimiLica?: number;
  godinaZastoyvane?: number;
  istoriyaNaUspeha?: string;
}

export interface EnterpriseUpdate {
  naimenovanie?: string;
  eik?: string;
  oblast?: string;
  grad?: string;
  adres?: string;
  lat?: number;
  lng?: number;
  klas?: string;
  pravnaForma?: string;
  ikonomicheskaDeynost?: string;
  celevnaGrupa?: string;
  socialnaKauza?: string;
  socialnaInovaciya?: string;
  kratkoOpisanie?: string;
  misiya?: string;
  telefon?: string;
  email?: string;
  website?: string;
  facebook?: string;
  logoUrl?: string;
  producti?: string;
  uslugi?: string;
  broyZaeti?: number;
  broyUyazvimiLica?: number;
  godinaZastoyvane?: number;
  istoriyaNaUspeha?: string;
}

export interface EnterpriseStats {
  totalEnterprises: number;
  totalZaeti: number;
  totalUyazvimiLica: number;
  totalInovacii: number;
  totalKauzi: number;
  totalSabytiya: number;
}

export interface OblastCount {
  oblast: string;
  count: number;
}

export interface SectorCount {
  sector: string;
  count: number;
}

export interface KauzaCount {
  kauza: string;
  count: number;
}

export interface Event {
  id: number;
  zaglavia: string;
  data: string;
  vid: string;
  opisanie: string;
  /** @nullable */
  myasto?: string | null;
  /** @nullable */
  organizator?: string | null;
  /** @nullable */
  linkRegistraciya?: string | null;
}

export interface EventInput {
  zaglavia: string;
  data: string;
  vid: string;
  opisanie: string;
  myasto?: string;
  organizator?: string;
  linkRegistraciya?: string;
}

export type ListEnterprisesParams = {
search?: string;
oblast?: string;
klas?: string;
pravnaForma?: string;
ikonomicheskaDeynost?: string;
celevGrupa?: string;
socialnaKauza?: string;
socialnaInovaciya?: string;
};