# lib/api-zod/src/generated/types/enterpriseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API за интерактивна карта на социалните предприятия в Североизточен район
 * OpenAPI spec version: 0.1.0
 */

export interface EnterpriseInput {
  naimenovanie: string;
  eik?: string;
  oblast: string;
  grad: string;
  adres?: string;
  lat: number;
  lng: number;
  klas: string;
  pravnaForma: string;
  ikonomicheskaDeynost: string;
  celevnaGrupa: string;
  socialnaKauza: string;
  socialnaInovaciya?: string;
  kratkoOpisanie: string;
  misiya?: string;
  telefon?: string;
  email?: string;
  website?: string;
  facebook?: string;
  logoUrl?: string;
  producti?: string;
  uslugi?: string;
  broyZaeti?: number;
  broyUyazvimiLica?: number;
  godinaZastoyvane?: number;
  istoriyaNaUspeha?: string;
}