# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API за интерактивна карта на социалните предприятия в Североизточен район
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  zaglavia: string;
  data: string;
  vid: string;
  opisanie: string;
  /** @nullable */
  myasto?: string | null;
  /** @nullable */
  organizator?: string | null;
  /** @nullable */
  linkRegistraciya?: string | null;
}