# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API за интерактивна карта на социалните предприятия в Североизточен район
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  zaglavia: string;
  data: string;
  vid: string;
  opisanie: string;
  myasto?: string;
  organizator?: string;
  linkRegistraciya?: string;
}