# lib/api-zod/src/generated/types/kauzaCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API за интерактивна карта на социалните предприятия в Североизточен район
 * OpenAPI spec version: 0.1.0
 */

export interface KauzaCount {
  kauza: string;
  count: number;
}